shelfaware/pyproject.toml
[tool.poetry]
name = "shelfaware"
version = "0.1.0"
description = "Backend for shelfaware meal planning and grocery inventory assistant"
authors = ["Matt Williamson <matt@aimatt.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
fastapi-cors = "^0.0.6"
uvicorn = "^0.31.0"
requests = "^2.32.3"
requests-cache = "^1.2.1"
matplotlib = "^3.9.2"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
pydantic = "^2.9.2"
ultralytics = "^8.3.5"
pillow = "^10.4.0"
opencv-python = "^4.10.0.84"
pyzbar = "^0.1.9"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
